ChangeList -Modified PagerAdapter to handle more fragment types -Previous RideAdapter properly formats the Day of the week according to the mocks -Fixed Error with ProfileActivity in uploading images -Fixed Error with Rating Drivers -Added swipe to refresh layout to wallet layout -Home Activity: --updated version shows driver position --shows only car icon once trip starts --updates user location every 10seconds --added helper functions to calculate distance between driver and rider --Added notifications for Driver proximity --Added a handler to monitor and calculate driver proximity --Added method to HomeActivityActions to hide nav menu and wallet menu once a trip has started and make them reappear on trip end --Proper handling of OnbackPressed by putting the app in stages to show each step of ordering a ride --Driver image is present --two methods for setting rider details depending on whether it's a fresh trip or the app is resuming from the backstack --An Event listener is added to the driver, once the driver is identified, this helps to monitor his position for proximity purposes --Implemented New Methods for the homeActivity -Changed Response Models to reflect changes in the api responses -Changed location updates time for GPStracker -Added new Icons to the app -Included the firebase db for the driver to the app for easier monitoring for proximity

+++ app/src/main/java/com/moovapp/riderapp/utils/retrofit/responseModels/VerifyPaymentResponseModel.java
@@ -45,13 +45,13 @@ public class VerifyPaymentResponseModel {
          * new_wallet_balance : 80
          */
 
-        private int new_wallet_balance;
+        private double new_wallet_balance;
 
-        public int getNew_wallet_balance() {
+        public double getNew_wallet_balance() {
             return new_wallet_balance;
         }
 
-        public void setNew_wallet_balance(int new_wallet_balance) {
+        public void setNew_wallet_balance(double new_wallet_balance) {
             this.new_wallet_balance = new_wallet_balance;
         }
     }
